Crosstalk Caffe Bug Fixes
Resolved bugs due to:
1. Scaling in batch normalization was done twice
2. Splicing did not take into account a default axis
3. Did not support newer protobuf versions.
4. Stdout was not being ignored during validation

+++ Source/CNTKv2LibraryDll/tensorboard/TensorBoardFileWriter.cpp
@@ -228,7 +228,7 @@ namespace CNTK
                     break;
 
                 default:
-                    fprintf(stderr, "TensorBoardFileWriter: Unsupported data type: %d ", dtype);
+                    fprintf(stderr, "TensorBoardFileWriter: Unsupported data type: %d ", static_cast<int>(dtype));
                     break;
                 }
